Remove unnecessary components
Removed ChangeDisplayNameButton & ChangePasswordButton out of their
components since they were basically wrapper classes with little
extra functionaliy (only router.push()) which can be done in the
AccountSettingsComponent so that only one router has to be created.
