+++ src/components/settings/account/account-settings.tsx
@@ -2,14 +2,26 @@ import React from "react"
 import SettingsWrapper from "@/components/settings/settings-wrapper"
 import { UserCircleIcon } from "@heroicons/react/24/outline"
 import SignOut from "@/components/settings/account/sign-out"
-import ChangePasswordButton from "@/components/settings/account/change-password-button"
-import ChangeDisplayNameButton from "@/components/settings/account/change-display-name-button"
+import SettingsButton from "@/components/settings/controls/settings-button"
+import { useRouter } from "next/router"
 
 const AccountSettings = () => {
+  const router = useRouter()
+
   return (
     <SettingsWrapper category={"Account"} icon={UserCircleIcon}>
-      <ChangeDisplayNameButton />
-      <ChangePasswordButton />
+      <SettingsButton
+        disabled={false}
+        onClick={async () => await router.push("/settings/change-display-name")}
+        text={"Change display name"}
+        dataCy={"change-display-name-button"}
+      />
+      <SettingsButton
+        disabled={false}
+        onClick={async () => await router.push("/settings/change-password")}
+        text={"Change password"}
+        dataCy={"change-password-button"}
+      />
       <SignOut />
     </SettingsWrapper>
   )
